Removed static type warning in GUISwarm. Reverted -handleButton argument types back to (int). --gepr
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@137 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/tkobjc/Raster.m
@@ -174,7 +174,7 @@
   return self;
 }
 
--handleButton: (int) n X: (long) x Y: (long) y {
+-handleButton: (int) n X: (int) x Y: (int) y {
   id c = 0;
   SEL sel = 0;
   switch(n) {
